replace app_registry with domain_ownership_service
