+++ app/lib/services/auth_service.rb
@@ -1,16 +1,30 @@
 module Services
   class AuthService
     def initialize
-      # TODO make this selectable
-      @impl = AppRegistryService.new
+      @domain_ownership_service = DomainOwnershipService.new
     end
 
     def authenticate!(token)
-      @impl.authenticate!(token)
+      identity = decode(token)
+      raise AuthError unless identity
+      # TODO verify identity with authority?
+      identity
     end
 
-    def authorize!(token, cert_issue_req)
-      @impl.authorize!(token, cert_issue_req)
+    def authorize!(identity, cert_issue_req)
+      @domain_ownership_service.authorize!(identity, cert_issue_req)
     end
+
+    private
+
+    def decode(token)
+      # Decode a JWT access token using the configured base.
+      body = JWT.decode(token, Rails.application.config.astral[:jwt_signing_key])[0]
+      Identity.new(body)
+    rescue => e
+      Rails.logger.warn "Unable to decode token: #{e}"
+      nil
+    end
+
   end
 end
