+++ app/models/domain_info.rb
@@ -1,0 +1,8 @@
+class DomainInfo
+  include ActiveModel::Model
+  include ActiveModel::Attributes
+
+  attribute :owner, :string
+  attribute :groups, array: :string, default: []
+  attribute :group_delegation, :bool, default: false
+end
