+++ app/models/identity.rb
@@ -1,0 +1,7 @@
+class Identity
+  include ActiveModel::Model
+  include ActiveModel::Attributes
+
+  attribute :subject, :string
+  attribute :groups, array: :string, default: []
+end
